fix(storage): read SQLite COUNT as i64 for rusqlite 0.32+ compatibility
rusqlite >=0.32 removed the FromSql impl for usize/isize. alert_counts and count_table read COUNT(*) directly into usize, which fails to compile on newer rusqlite. Read as i64 (SQLite's integer type) and cast to usize. Compatible with the current 0.31 pin and unblocks the rusqlite 0.40 bump in #116.

+++ src/storage.rs
@@ -709,8 +709,9 @@ impl StorageBackend {
             let _ = stmt
                 .query_map(params_ref.as_slice(), |row| {
                     let level: String = row.get(0)?;
-                    let count: usize = row.get(1)?;
-                    Ok((level, count))
+                    // rusqlite >=0.32 dropped FromSql for usize; SQLite COUNT(*) is i64.
+                    let count: i64 = row.get(1)?;
+                    Ok((level, count as usize))
                 })
                 .map(|rows| {
                     for r in rows.flatten() {
@@ -1220,8 +1221,9 @@ impl StorageBackend {
         // table names are hardcoded constants, not user input
         let sql = format!("SELECT COUNT(*) FROM {table}");
         self.conn
-            .query_row(&sql, [], |row| row.get::<_, usize>(0))
-            .unwrap_or(0)
+            // rusqlite >=0.32 dropped FromSql for usize; SQLite COUNT(*) returns i64.
+            .query_row(&sql, [], |row| row.get::<_, i64>(0))
+            .unwrap_or(0) as usize
     }
 
     /// Get storage statistics.
